backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0

zai-sdk==0.1.0
sniffio>=1.3.0
httpx>=0.27.0
anyio>=4.0.0

pillow>=10.0.0
python-dotenv>=1.0.0
